done project and init test all routes

+++ app/Helpers/ClassesBase/Routes/CrudActions.php
@@ -9,6 +9,8 @@ use App\Helpers\ClassesProcess\RolesPermissions\PermissionsMap;
 
 abstract class CrudActions
 {
+    public bool $isActive = true;
+
     protected ?RouteAction $indexAction = null;
     protected ?RouteAction $indexTrashesAction = null;
     protected ?RouteAction $indexAllAction = null;
@@ -48,27 +50,29 @@ abstract class CrudActions
     protected abstract function controller():string;
 
     private function initActions(){
-        $this->indexAction = new RouteAction("","index","index","get",[$this->getPermissionWithNameCrud(PermissionsMap::READ),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->indexTrashesAction = new RouteAction("trashes","index.trashes","indexTrashes","get",[$this->getPermissionWithNameCrud(PermissionsMap::READ_TRASHES),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->indexAllAction = new RouteAction("all","get.all","indexAll","get",[$this->getPermissionWithNameCrud(PermissionsMap::READ),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->showSearchFieldsAction = new RouteAction("get/fields/show_search","get.fields.show_search","showSearchFields","get",[$this->getPermissionWithNameCrud(PermissionsMap::READ),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->createAction = new RouteAction("create","create","create","get",[$this->getPermissionWithNameCrud(PermissionsMap::CREATE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->storeAction = new RouteAction("create","store","store","post",[$this->getPermissionWithNameCrud(PermissionsMap::CREATE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->showAction = new RouteAction("show/{id}","show","show","get",[$this->getPermissionWithNameCrud(PermissionsMap::READ),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
-        $this->editAction = new RouteAction("edit/{id}","edit","edit","get",[$this->getPermissionWithNameCrud(PermissionsMap::UPDATE),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
-        $this->updateAction = new RouteAction("edit/{id}","update","update","put",[$this->getPermissionWithNameCrud(PermissionsMap::UPDATE),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
-        $this->deleteAction = new RouteAction("delete/{id}","delete","delete","delete",[$this->getPermissionWithNameCrud(PermissionsMap::DELETE),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
-        $this->deleteMultiAction = new RouteAction("delete/records","delete.records","multiDelete","delete",[$this->getPermissionWithNameCrud(PermissionsMap::DELETE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->forceDeleteAction = new RouteAction("force/delete/{id}","force.delete","forceDelete","delete",[$this->getPermissionWithNameCrud(PermissionsMap::FORCE_DELETE),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
-        $this->forceDeleteMultiAction = new RouteAction("force/delete/records","force.delete.records","multiForceDelete","delete",[$this->getPermissionWithNameCrud(PermissionsMap::FORCE_DELETE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->restoreAction = new RouteAction("restore/{id}","restore","restore","post",[$this->getPermissionWithNameCrud(PermissionsMap::RESTORE),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
-        $this->restoreMultiAction = new RouteAction("restore/records","restore.records","multiRestore","post",[$this->getPermissionWithNameCrud(PermissionsMap::RESTORE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->restoreAllAction = new RouteAction("restore/all/records","restore.all.records","restoreAll","post",[$this->getPermissionWithNameCrud(PermissionsMap::RESTORE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->clearTrashesAction = new RouteAction("clear/trashes","clear.trashes","clearTrashes","delete",[$this->getPermissionWithNameCrud(PermissionsMap::FORCE_DELETE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->exportXLSXAction = new RouteAction("export/excel","export.excel","exportXLSX","get",[$this->getPermissionWithNameCrud(PermissionsMap::Export),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
-        $this->exportPDFAction = new RouteAction("export/pdf","export.pdf","exportPDF","get",[$this->getPermissionWithNameCrud(PermissionsMap::Export),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+        if($this->isActive){
+            $this->indexAction = new RouteAction("","index","index","get",[$this->getPermissionWithNameCrud(PermissionsMap::READ),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->indexTrashesAction = new RouteAction("trashes","index.trashes","indexTrashes","get",[$this->getPermissionWithNameCrud(PermissionsMap::READ_TRASHES),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->indexAllAction = new RouteAction("all","get.all","indexAll","get",[$this->getPermissionWithNameCrud(PermissionsMap::READ),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->showSearchFieldsAction = new RouteAction("get/fields/show_search","get.fields.show_search","showSearchFields","get",[$this->getPermissionWithNameCrud(PermissionsMap::READ),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->createAction = new RouteAction("create","create","create","get",[$this->getPermissionWithNameCrud(PermissionsMap::CREATE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->storeAction = new RouteAction("create","store","store","post",[$this->getPermissionWithNameCrud(PermissionsMap::CREATE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->showAction = new RouteAction("show/{id}","show","show","get",[$this->getPermissionWithNameCrud(PermissionsMap::READ),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
+            $this->editAction = new RouteAction("edit/{id}","edit","edit","get",[$this->getPermissionWithNameCrud(PermissionsMap::UPDATE),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
+            $this->updateAction = new RouteAction("edit/{id}","update","update","put",[$this->getPermissionWithNameCrud(PermissionsMap::UPDATE),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
+            $this->deleteAction = new RouteAction("delete/{id}","delete","delete","delete",[$this->getPermissionWithNameCrud(PermissionsMap::DELETE),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
+            $this->deleteMultiAction = new RouteAction("delete/records","delete.records","multiDelete","delete",[$this->getPermissionWithNameCrud(PermissionsMap::DELETE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->forceDeleteAction = new RouteAction("force/delete/{id}","force.delete","forceDelete","delete",[$this->getPermissionWithNameCrud(PermissionsMap::FORCE_DELETE),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
+            $this->forceDeleteMultiAction = new RouteAction("force/delete/records","force.delete.records","multiForceDelete","delete",[$this->getPermissionWithNameCrud(PermissionsMap::FORCE_DELETE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->restoreAction = new RouteAction("restore/{id}","restore","restore","post",[$this->getPermissionWithNameCrud(PermissionsMap::RESTORE),$this->getPermissionWithNameCrud(PermissionsMap::All)],[],"id");
+            $this->restoreMultiAction = new RouteAction("restore/records","restore.records","multiRestore","post",[$this->getPermissionWithNameCrud(PermissionsMap::RESTORE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->restoreAllAction = new RouteAction("restore/all/records","restore.all.records","restoreAll","post",[$this->getPermissionWithNameCrud(PermissionsMap::RESTORE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->clearTrashesAction = new RouteAction("clear/trashes","clear.trashes","clearTrashes","delete",[$this->getPermissionWithNameCrud(PermissionsMap::FORCE_DELETE),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->exportXLSXAction = new RouteAction("export/excel","export.excel","exportXLSX","get",[$this->getPermissionWithNameCrud(PermissionsMap::Export),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->exportPDFAction = new RouteAction("export/pdf","export.pdf","exportPDF","get",[$this->getPermissionWithNameCrud(PermissionsMap::Export),$this->getPermissionWithNameCrud(PermissionsMap::All)]);
+            $this->actions = Arr::except(get_object_vars($this),["actions","nameCrud","middlewares","repository","finalData","isActive"]);
+        }
         $this->handle();
-        $this->actions = Arr::except(get_object_vars($this),["actions","nameCrud","middlewares","repository","finalData"]);
     }
 
     private function getPermissionWithNameCrud(string $permission): string
